move GameAnalyticsEvent wire codec off the domain ADT to the edge
