+++ game-service-server/src/main/scala/com/andy327/server/analytics/GameAnalyticsEvent.scala
@@ -1,15 +1,12 @@
 package com.andy327.server.analytics
 
-import io.circe.syntax._
-import io.circe.{Decoder, Encoder, Json}
-
 import com.andy327.model.core.{GameId, GameType, PlayerId}
-import com.andy327.persistence.db.schema.GameTypeCodecs.gameTypeCodec
 
 /** Structured domain events describing a game's lifecycle, emitted from the actor layer for analytics. Unlike
   * [[com.andy327.server.actors.core.PlayerEvent]] (which carries rendered board state for WebSocket delivery), these
   * carry only the dimensions analytics needs. Published on the `game-analytics` channel and decoded by
-  * [[AnalyticsConsumer]], so the codec must round-trip.
+  * [[AnalyticsConsumer]]; the wire codec lives in [[GameAnalyticsCodecs]], keeping this ADT free of any serialization
+  * concern.
   */
 sealed trait GameAnalyticsEvent {
 
@@ -37,10 +34,6 @@ object GameAnalyticsEvent {
     val all: List[Outcome] = List(Won, Draw, Forfeit)
 
     def fromLabel(s: String): Option[Outcome] = all.find(_.label == s)
-
-    implicit val encoder: Encoder[Outcome] = Encoder.encodeString.contramap(_.label)
-    implicit val decoder: Decoder[Outcome] =
-      Decoder.decodeString.emap(s => fromLabel(s).toRight(s"Unknown Outcome: $s"))
   }
 
   /** A new game was created and started. */
@@ -55,71 +48,4 @@ object GameAnalyticsEvent {
 
   /** A chat message was sent. `gameType` is `None` for a message sent in a lobby before the game starts. */
   final case class ChatSent(gameId: GameId, gameType: Option[GameType]) extends GameAnalyticsEvent
-
-  /** Each variant is encoded as a JSON object with a `type` discriminator, mirroring the convention used by
-    * [[com.andy327.server.http.json.JsonProtocol.playerEventEncoder]].
-    */
-  implicit val encoder: Encoder[GameAnalyticsEvent] = Encoder.instance {
-    case GameStarted(gameId, gameType, playerCount) =>
-      Json.obj(
-        "type" -> "GameStarted".asJson,
-        "gameId" -> gameId.asJson,
-        "gameType" -> gameType.asJson,
-        "playerCount" -> playerCount.asJson
-      )
-    case MoveMade(gameId, gameType, playerId, seq) =>
-      Json.obj(
-        "type" -> "MoveMade".asJson,
-        "gameId" -> gameId.asJson,
-        "gameType" -> gameType.asJson,
-        "playerId" -> playerId.asJson,
-        "seq" -> seq.asJson
-      )
-    case GameCompleted(gameId, gameType, outcome, moveCount) =>
-      Json.obj(
-        "type" -> "GameCompleted".asJson,
-        "gameId" -> gameId.asJson,
-        "gameType" -> gameType.asJson,
-        "outcome" -> outcome.asJson,
-        "moveCount" -> moveCount.asJson
-      )
-    case ChatSent(gameId, gameType) =>
-      Json.obj(
-        "type" -> "ChatSent".asJson,
-        "gameId" -> gameId.asJson,
-        "gameType" -> gameType.asJson
-      )
-  }
-
-  implicit val decoder: Decoder[GameAnalyticsEvent] = Decoder.instance { c =>
-    c.get[String]("type").flatMap {
-      case "GameStarted" =>
-        for {
-          gameId <- c.get[GameId]("gameId")
-          gameType <- c.get[GameType]("gameType")
-          playerCount <- c.get[Int]("playerCount")
-        } yield GameStarted(gameId, gameType, playerCount)
-      case "MoveMade" =>
-        for {
-          gameId <- c.get[GameId]("gameId")
-          gameType <- c.get[GameType]("gameType")
-          playerId <- c.get[PlayerId]("playerId")
-          seq <- c.get[Int]("seq")
-        } yield MoveMade(gameId, gameType, playerId, seq)
-      case "GameCompleted" =>
-        for {
-          gameId <- c.get[GameId]("gameId")
-          gameType <- c.get[GameType]("gameType")
-          outcome <- c.get[Outcome]("outcome")
-          moveCount <- c.get[Int]("moveCount")
-        } yield GameCompleted(gameId, gameType, outcome, moveCount)
-      case "ChatSent" =>
-        for {
-          gameId <- c.get[GameId]("gameId")
-          gameType <- c.get[Option[GameType]]("gameType")
-        } yield ChatSent(gameId, gameType)
-      case other =>
-        Left(io.circe.DecodingFailure(s"Unknown GameAnalyticsEvent type: $other", c.history))
-    }
-  }
 }
